+++ web/index.jade
@@ -11,7 +11,7 @@ html
 
           // All paths must be same origin.
 
-          // Array of CSS file paths.
+          // Array of CSS file paths to document.
           stylesheets: [
             // 'stylesheet.css',
           ],
